JavaScript (v3): CloudWatch Logs Wkflw - Rename 'big query' to 'log query' to remove any confusion with GCP.
