+++ .doc_gen/metadata/cloudwatch-logs_metadata.yaml
@@ -348,7 +348,7 @@ cloudwatch-logs_Scenario_BigQuery:
                 The first file is the entry point. The second file is a class
                 that splits queries into multiple steps if necessary.
               snippet_files:
-                - javascriptv3/example_code/cloudwatch-logs/scenarios/big-query/index.js
-                - javascriptv3/example_code/cloudwatch-logs/scenarios/big-query/cloud-watch-query.js
+                - javascriptv3/example_code/cloudwatch-logs/scenarios/large-query/index.js
+                - javascriptv3/example_code/cloudwatch-logs/scenarios/large-query/cloud-watch-query.js
   services:
     cloudwatch-logs: {StartQuery, GetQueryResults}
