+++ javascriptv3/example_code/cloudwatch-logs/scenarios/large-query/index.js
