+++ javascriptv3/example_code/cloudwatch-logs/tests/cloudwatch-query.unit.test.js
@@ -1,5 +1,5 @@
 import { describe, it, vi, expect } from "vitest";
-import { CloudWatchQuery } from "../scenarios/big-query/cloud-watch-query.js";
+import { CloudWatchQuery } from "../scenarios/large-query/cloud-watch-query.js";
 import {
   DescribeQueriesCommand,
   GetQueryResultsCommand,
