--- workflows/cloudwatch_logs_big_query/resources/stack.yaml
@@ -1,10 +1,0 @@
-Resources:
-  BigQueryLogGroup:
-    Type: AWS::Logs::LogGroup
-    Properties:
-      LogGroupName: /workflows/cloudwatch-log/big-query
-  BigQueryLogGroupStream1:
-    Type: AWS::Logs::LogStream
-    Properties:
-      LogGroupName: !Ref BigQueryLogGroup
-      LogStreamName: stream1
