+++ workflows/cloudwatch_logs_large_query/resources/make-log-files.sh
