+++ workflows/cloudwatch_logs_large_query/resources/put-log-events.sh
@@ -8,9 +8,9 @@ for f in $FILES; do
     LOG_GROUP_NAME=""
     
     if [[ "$OSTYPE" == "msys" ]]; then
-        LOG_GROUP_NAME="//workflows\\cloudwatch-log\\big-query"
+        LOG_GROUP_NAME="//workflows\\cloudwatch-logs\\large-query"
     else
-        LOG_GROUP_NAME="/workflows/cloudwatch-log/big-query"
+        LOG_GROUP_NAME="/workflows/cloudwatch-logs/large-query"
     fi
 
     aws logs put-log-events --log-group-name "$LOG_GROUP_NAME" --log-stream-name stream1 --log-events file://"$f" --no-cli-pager
